pyproject.toml
[project]
name = "ai-agent"
description = "General-purpose AI agent with MCP tool integration"
version = "0.1.0"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pydantic>=2.11.5",
    "bubus>=1.4.7",
    "mcp>=1.10.1",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.1",
    "typing-extensions>=4.12.2",
    "aiofiles>=24.1.0",
    "anyio>=4.9.0",
    "anthropic>=0.58.2",
    "openai>=1.97.0",
    "google-genai>=1.26.0",
    "groq>=0.30.0",
    "ollama>=0.5.1",
    "uuid7>=0.1.0"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.11.2",
    "pyright>=1.1.403",
    "pytest>=8.3.5",
    "pytest-asyncio>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 130
fix = true

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501", "F841"]

[tool.ruff.format]
quote-style = "single"
indent-style = "tab"
line-ending = "lf"

[tool.pyright]
typeCheckingMode = "basic"
exclude = [".venv/", ".git/", "__pycache__/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]